backend/api-gateway/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
cryptography==43.0.1
httpx==0.27.2
python-dotenv==1.0.1

